test: make ear more complex to test classloading
