+++ dev/com.ibm.ws.concurrent_fat_cdi/test-applications/concurrentCDIApp/resources/META-INF/application.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-    Copyright (c) 2024 IBM Corporation and others.
+    Copyright (c) 2024, 2025 IBM Corporation and others.
     All rights reserved. This program and the accompanying materials
     are made available under the terms of the Eclipse Public License 2.0
     which accompanies this distribution, and is available at
@@ -40,9 +40,9 @@
 
   <context-service>
     <name>java:app/concurrent/with-location-and-without-app-context</name>
-    <qualifier>concurrent.cdi.ejb.ClearingAppContext</qualifier>
-    <qualifier>concurrent.cdi.ejb.PropagatingLocationContext</qualifier>
-    <qualifier>concurrent.cdi.ejb.IgnoringTransactionContext</qualifier>
+    <qualifier>concurrent.cdi.ejb.anno.ClearingAppContext</qualifier>
+    <qualifier>concurrent.cdi.ejb.anno.PropagatingLocationContext</qualifier>
+    <qualifier>concurrent.cdi.ejb.anno.IgnoringTransactionContext</qualifier>
     <cleared>Application</cleared>
     <propagated>Location</propagated>
     <unchanged>Remaining</unchanged>
@@ -50,14 +50,14 @@
 
   <managed-thread-factory>
     <name>java:app/concurrent/thread-factory-app-dd-with-app-context</name>
-    <qualifier>concurrent.cdi.ejb.PropagatingAppContext</qualifier>
+    <qualifier>concurrent.cdi.ejb.anno.PropagatingAppContext</qualifier>
     <context-service-ref>java:app/concurrent/with-app-context</context-service-ref>
     <priority>1</priority>
   </managed-thread-factory>
 
   <managed-thread-factory>
     <name>java:app/concurrent/thread-factory-app-dd-without-app-context</name>
-    <qualifier>concurrent.cdi.ejb.ClearingAppContext</qualifier>
+    <qualifier>concurrent.cdi.ejb.anno.ClearingAppContext</qualifier>
     <context-service-ref>java:app/concurrent/with-location-and-tx-context</context-service-ref>
     <priority>2</priority>
   </managed-thread-factory>
